Change some variables from float to uint32_t

+++ src/FSM/Inc/Apps.h
@@ -31,12 +31,12 @@
 
 // Fault handling variables
 #define PEDAL_SATURATION_TIMEOUT  (uint32_t)(10)         // (sec)
-#define APPS_IMPLAUSIBILITY_DELAY (float32_t)(1.0f)       // (sec)
-#define APPS_BPPC_THRESHOLD       (float32_t)(0.25f)      // (decimal %)
+#define APPS_IMPLAUSIBILITY_DELAY (uint32_t)(1)          // (sec)
+#define APPS_BPPC_THRESHOLD       (float32_t)(0.25f)     // (decimal %)
 #define MAX_APPS_FAULTS \
-    (float32_t)APPS_IMPLAUSIBILITY_DELAY * (float32_t)CONTROL_LOOP_FREQUENCY
+    (uint32_t)(APPS_IMPLAUSIBILITY_DELAY * CONTROL_LOOP_FREQUENCY)
 #define MAX_SATURATION_FAULTS \
-    (uint32_t)PEDAL_SATURATION_TIMEOUT * (uint32_t)CONTROL_LOOP_FREQUENCY
+    (uint32_t)(PEDAL_SATURATION_TIMEOUT * CONTROL_LOOP_FREQUENCY)
 
 /******************************************************************************
  * Preprocessor Macros
